Change default RNumber to Float

+++ src/structures/RNumber.js
@@ -17,7 +17,7 @@ class RNumber {
 
   static parse(text) {
     let amount = null;
-    let type = RNumber.types.INT;
+    let type = RNumber.types.FLOAT;
     // Remove mentions
     text = text.replace(/<.{2}\d+>/i, "").trim();
     // Get the first number in the text
